[Refactor] 홈뷰 피드백 반영
