+++ mohaemukzip/Sources/Feature/Home/Views/MonthlyCookingLineChart.swift
@@ -1,10 +1,3 @@
-//
-//  MonthlyCookingLineChart.swift
-//  mohaemukzip
-//
-//  Created by 이서현 on 1/28/26.
-//
-
 
 import SwiftUI
 import Charts
@@ -22,7 +15,6 @@ struct MonthlyCookingLineChart: View {
         self.points = points
     }
     
-    /// Preview / placeholder initializer
     init() {
         self.points = [
             .init(month: 1, count: 2),
@@ -64,13 +56,12 @@ struct MonthlyCookingLineChart: View {
             }
         }
         .chartXScale(domain: 1...12)
-        .chartXScale(range: .plotDimension(padding: 16))
         .chartYScale(domain: 0...maxY + 1)
         .chartYAxis(.hidden)
         .chartLegend(.hidden)
         .chartXAxis {
-            AxisMarks(values: Array(1...12)) { value in
-                AxisValueLabel {
+            AxisMarks(values: Array(stride(from: 1, through: 12, by: 1))) { value in
+                AxisValueLabel() {
                     if let m = value.as(Int.self) {
                         Text("\(m)월")
                             .font(.PretendardRegular13)
@@ -80,13 +71,17 @@ struct MonthlyCookingLineChart: View {
             }
         }
         .chartPlotStyle { plotArea in
-            plotArea.background(Color.clear)
+            plotArea
+                .padding(.horizontal, 16)
+                .background(Color.clear)
+        }
+        // Charts가 공간 부족 시 마지막 축 라벨(12월)을 자동 생략하는 경우가 있어서
+        // 12월만 수동으로 한 번 더 찍어줬습니다
+        .overlay(alignment: .bottomTrailing) {
+            Text("12월")
+                .font(.PretendardRegular13)
+                .foregroundStyle(.grey700)
+                .padding(.trailing, 4)
         }
     }
-}
-
-
-
-#Preview {
-    MonthlyCookingLineChart()
 }
